fix: optimize repost logic

+++ lib/app/features/feed/reposts/providers/optimistic/repost_sync_strategy.dart
@@ -23,15 +23,17 @@ class RepostSyncStrategy implements SyncStrategy<PostRepost> {
     if (toggledToRepost) {
       final createdRepostReference = await createRepost(optimistic.eventReference);
       invalidateCounterCache(optimistic.eventReference);
-
-      return optimistic.copyWith(myRepostReference: createdRepostReference);
-    } else if (toggledToUnrepost) {
-      if (previous.myRepostReference != null) {
-        await deleteRepost(previous.myRepostReference!);
-        invalidateCounterCache(optimistic.eventReference);
-      }
-
-      return optimistic;
+      
+      return optimistic.copyWith(
+        myRepostReference: createdRepostReference,
+      );
+    } else if (toggledToUnrepost && previous.myRepostReference != null) {
+      await deleteRepost(previous.myRepostReference!);
+      invalidateCounterCache(optimistic.eventReference);
+      
+      return optimistic.copyWith(
+        myRepostReference: null,
+      );
     }
 
     return optimistic;
